Restructure the logic
