+++ datamatic/generator.py
@@ -1,5 +1,4 @@
 import re
-import pathlib
 from typing import Tuple, Literal, Optional
 from dataclasses import dataclass, field
 from functools import partial
